Move footer out of content so it can be extended properly
If we wrap the exsting <footer> inside <context> in a {%- block footer %} and
then extend this, it will be represented twice in the final rendered document.
This is because in the original basic/layout.html, {%- block footer %} is
defined outside of {%- block content %}. So when the content-block is rendered
it will pull in sphinx-material's content-block (containing the footer-block)
and then later the footer-block will be rendered again.

To avoid this, this commit pulls the footer-block out of the content-block,
matching the structure of the parent template basic/layout.html. This does
depart from how mkdocs-material does things, however I believe that this is
also a bug on their end, as their <header> is defined outside of content but
their <footer> is defined inside of content, which is a strange inconsistency.

The new behaviour doesn't seem to have any adverse effects, such as CSS rules
or anything else depending on footer being inside content previously.